# [2/3] Remove computed GOTOs