--- a/src/arm64/implementation.cpp
+++ b/src/arm64/implementation.cpp
@@ -17,6 +17,6 @@ WARN_UNUSED error_code implementation::create_dom_parser_implementation(
 }
 
 } // namespace SIMDJSON_IMPLEMENTATION
-} // namespace {
+} // unnamed namespace
 
 #include "arm64/end_implementation.h"