--- a/src/fallback/bitmanipulation.h
+++ b/src/fallback/bitmanipulation.h
@@ -24,9 +24,10 @@ static unsigned char _BitScanReverse64(unsigned long* ret, uint64_t x) {
 }
 #endif
 
-// We sometimes call trailing_zero on inputs that are zero,
-// but the algorithms do not end up using the returned value.
-// Sadly, sanitizers are not smart enough to figure it out.
+//
+// These are currently unused, but one day will be.
+//
+#if 0 // Currently unused
 NO_SANITIZE_UNDEFINED
 really_inline int trailing_zeroes(uint64_t input_num) {
 #ifdef _MSC_VER
@@ -44,6 +45,7 @@ really_inline int trailing_zeroes(uint64_t input_num) {
 really_inline uint64_t clear_lowest_bit(uint64_t input_num) {
   return input_num & (input_num-1);
 }
+#endif // Currently unused
 
 /* result might be undefined when input_num is zero */
 really_inline int leading_zeroes(uint64_t input_num) {
@@ -60,6 +62,7 @@ really_inline int leading_zeroes(uint64_t input_num) {
 #endif// _MSC_VER
 }
 
+#if 0 // Currently unused
 really_inline bool add_overflow(uint64_t value1, uint64_t value2, uint64_t *result) {
   *result = value1 + value2;
   return *result < value1;
@@ -70,6 +73,7 @@ really_inline bool mul_overflow(uint64_t value1, uint64_t value2, uint64_t *resu
   // TODO there must be a faster way
   return value2 > 0 && value1 > std::numeric_limits<uint64_t>::max() / value2;
 }
+#endif // Currently unused
 
 } // namespace fallback
 } // namespace {