--- a/src/fallback/implementation.cpp
+++ b/src/fallback/implementation.cpp
@@ -17,6 +17,6 @@ WARN_UNUSED error_code implementation::create_dom_parser_implementation(
 }
 
 } // namespace SIMDJSON_IMPLEMENTATION
-} // namespace {
+} // unnamed namespace
 
 #include "fallback/end_implementation.h"