--- a/src/fallback/implementation.h
+++ b/src/fallback/implementation.h
@@ -27,6 +27,6 @@ public:
 };
 
 } // namespace fallback
-} // namespace {
+} // unnamed namespace
 
 #endif // SIMDJSON_FALLBACK_IMPLEMENTATION_H