--- a/src/fallback/numberparsing.h
+++ b/src/fallback/numberparsing.h
@@ -22,7 +22,7 @@ static really_inline uint32_t parse_eight_digits_unrolled(const uint8_t *chars) 
 }
 
 } // namespace SIMDJSON_IMPLEMENTATION
-} // namespace {
+} // unnamed namespace
 
 #define SWAR_NUMBER_PARSING
 #include "generic/stage2/numberparsing.h"