--- a/src/fallback/stringparsing.h
+++ b/src/fallback/stringparsing.h
@@ -3,9 +3,8 @@
 
 #include "simdjson.h"
 
-namespace simdjson {
-namespace fallback {
 namespace {
+namespace fallback {
 
 // Holds backslashes and quotes locations.
 struct backslash_and_quote {
@@ -27,9 +26,8 @@ really_inline backslash_and_quote backslash_and_quote::copy_and_find(const uint8
   return { src[0] };
 }
 
-} // namespace {
 } // namespace fallback
-} // namespace simdjson
+} // namespace {
 
 #include "generic/stage2/stringparsing.h"