--- a/src/fallback/stringparsing.h
+++ b/src/fallback/stringparsing.h
@@ -27,7 +27,7 @@ really_inline backslash_and_quote backslash_and_quote::copy_and_find(const uint8
 }
 
 } // namespace fallback
-} // namespace {
+} // unnamed namespace
 
 #include "generic/stage2/stringparsing.h"