Second in a three-part PR series: #1065 (anonymous namespaces) -- 1078 (remove computed GOTOs) -- #1079 (return errors instead of goto).

This:
* Replaces all computed gotos with static gotos, changing actual flow as minimally as possible.
* Removes unified_machine_addresses, ret_address_t, and parse_value(), simplifying the parse_value switch to keep the main loop as simple as possible.
* Replaces the stack of continue addresses with a stack of booleans indicating whether each open scope is an object or an array.
* Marks parse_structurals() (the method with the gotos) as really_inline, indicating the inlining infrastructure may be able to work with our code better now that we're not using dynamic gotos.

The primary goal is to simplify control flow and use fewer fancy features where we can. This is also a necessary step towards a clean SAX interface, separating stage 2 iteration from stage 2 write. Otherwise, callbacks will have to receive (and return) computed goto addresses to keep the same level of performance as simdjson's current stage 2.

## Performance

This is a minor loss on Intel. It fluctuates each time I make a change. Instruction count is almost randomly all over the place. I think it's jitters. I have no idea how to prove otherwise.

ARM gcc10 impact is a very minor loss for some files, but doesn't make anything worse. Instruction count is all over the place, just like with Intel.

This does not appear to make Windows much faster (less than 1% for twitter.json on my machine, which is admittedly a laptop).

### skylake-x gcc9

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |
| mesh.pretty                      |  24646 | 184.7 | 180.3 |    2% | 584.1 | 579.2 |    0% |
| canada                           |  35172 | 295.9 | 289.1 |    2% | 982.0 | 982.2 |    0% |
| twitterescaped                   |   8787 | 190.8 | 190.5 |    0% | 512.0 | 515.5 |    0% |
| instruments                      |   3442 | 108.5 | 108.8 |    0% | 376.1 | 373.6 |    0% |
| twitter                          |   9867 |  91.6 |  92.6 |   -1% | 296.9 | 296.8 |    0% |
| citm_catalog                     |  26987 |  83.6 |  84.8 |   -1% | 312.5 | 314.5 |    0% |
| github_events                    |   1017 |  77.3 |  78.5 |   -1% | 267.3 | 269.4 |    0% |
| gsoc-2018                        |  51997 |  72.9 |  74.2 |   -1% | 169.0 | 169.8 |    0% |
| mesh                             |  11306 | 304.7 | 311.2 |   -2% | 989.7 | 978.9 |    1% |
| random                           |   7976 | 141.6 | 145.4 |   -2% | 496.8 | 497.3 |    0% |
| marine_ik                        |  46616 | 291.0 | 299.2 |   -2% | 915.6 | 912.6 |    0% |
| update-center                    |   8330 | 113.0 | 116.6 |   -3% | 342.6 | 344.8 |    0% |
| apache_builds                    |   1988 |  84.3 |  87.8 |   -4% | 307.4 | 311.4 |   -1% |
| numbers                          |   2345 | 241.1 | 253.4 |   -5% | 799.3 | 790.8 |    1% |

There is a great deal of apparent compiler-generated randomness here. For example, 

* [Make parse_structurals really_inline](https://github.com/simdjson/simdjson/pull/1078/commits/460cfcaf3e42fb86bc107dcc7c080dfeb71520ed) makes it jump around randomly but generally reduces instruction count:

  | File                             | Blocks | +Throughput | -Instr. |
  | ---                              |    --: |   --: |   --: |
  | marine_ik                        |  46616 |    4% |    0% |
  | numbers                          |   2345 |    4% |    0% |
  | citm_catalog                     |  26987 |    3% |    2% |
  | mesh                             |  11306 |    3% |    1% |
  | mesh.pretty                      |  24646 |    2% |    0% |
  | canada                           |  35172 |    0% |    0% |
  | twitterescaped                   |   8787 |    0% |    0% |
  | apache_builds                    |   1988 |    0% |    0% |
  | update-center                    |   8330 |    0% |    0% |
  | gsoc-2018                        |  51997 |   -1% |    0% |
  | github_events                    |   1017 |   -1% |    0% |
  | random                           |   7976 |   -1% |    0% |
  | twitter                          |   9867 |   -2% |    0% |
  | instruments                      |   3442 |   -5% |    0% |

### ARM gcc9

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| canada                           |  35172 | 852.6 | 837.5 |    1% | 1018.3 | 1026.4 |    0% |  52366 |  57395 |   -9% |  91681 |  91733 |    0% | 4904465 | 5103335 |   -4% |
| citm_catalog                     |  26987 | 397.7 | 393.1 |    1% | 370.6 | 373.4 |    0% |   4841 |   4702 |    2% |  62736 |  62727 |    0% | 1924043 | 1959557 |   -1% |
| mesh                             |  11306 | 949.4 | 940.4 |    0% | 1062.2 | 1054.9 |    0% |  21073 |  20218 |    4% |  32325 |  32344 |    0% | 1622781 | 1683179 |   -3% |
| marine_ik                        |  46616 | 893.6 | 886.1 |    0% | 988.2 | 987.6 |    0% |  99457 | 100457 |   -1% | 134021 | 133966 |    0% | 7439148 | 7781571 |   -4% |
| mesh.pretty                      |  24646 | 597.5 | 593.5 |    0% | 635.0 | 631.6 |    0% |  39558 |  40186 |   -1% |  59084 |  59080 |    0% | 1971802 | 2032827 |   -3% |
| instruments                      |   3442 | 482.0 | 479.5 |    0% | 462.8 | 462.6 |    0% |   3556 |   3215 |    9% |   8696 |   8674 |    0% | 335597 | 338701 |    0% |
| apache_builds                    |   1988 | 472.7 | 470.2 |    0% | 414.3 | 418.2 |    0% |    657 |    628 |    4% |   4836 |   4832 |    0% | 201636 | 203258 |    0% |
| gsoc-2018                        |  51997 | 360.1 | 360.3 |    0% | 238.8 | 239.4 |    0% |  38814 |  39638 |   -2% | 109478 | 109696 |    0% | 2755135 | 2763570 |    0% |
| twitter                          |   9867 | 489.4 | 489.7 |    0% | 386.9 | 387.2 |    0% |  16007 |  15477 |    3% |  23437 |  23435 |    0% | 871899 | 871078 |    0% |
| twitterescaped                   |   8787 | 736.7 | 737.3 |    0% | 629.4 | 629.8 |    0% |  19334 |  17907 |    7% |  21356 |  21356 |    0% | 1574717 | 1572639 |    0% |
| update-center                    |   8330 | 546.5 | 547.8 |    0% | 463.1 | 465.8 |    0% |  13723 |  13289 |    3% |  20804 |  20805 |    0% | 1037169 | 1033504 |    0% |
| random                           |   7976 | 701.8 | 703.7 |    0% | 639.4 | 643.1 |    0% |   8593 |   6693 |   22% |  21618 |  21609 |    0% | 1250791 | 1249231 |    0% |
| numbers                          |   2345 | 907.1 | 909.8 |    0% | 874.0 | 869.7 |    0% |   3283 |   3268 |    0% |   6030 |   6015 |    0% | 239948 | 259502 |   -8% |
| github_events                    |   1017 | 460.6 | 462.2 |    0% | 358.8 | 359.9 |    0% |   1523 |   1553 |   -1% |   2412 |   2406 |    0% |  88367 |  88624 |    0% |