@lemire 2020-07-30
@jkeiser Odd that it would not help Visual Studio.