@lemire 2020-07-30
Compared to master, my own tests do not indicate a performance regression.