@lemire 2020-07-30
One reason why this PR works well performance-wise is that the new switch/case constructs are compiled systematically into tight lookup tables...

https://godbolt.org/z/53MTGY

The clang compiler seems to do the best job with it but all compilers appear to deal with this code nicely.