@lemire 2020-07-31
I have converted the PR to a draft because I want @jkeiser  to consider adding an "unlikely" keyword on one conditional. It could change the performance outlook, at least for some cases.