@jkeiser 2020-08-02
@lemire I added the two `unlikely` branches and removed the ` else { }` from `else { goto }`, and this is now a definite improvement against master (updated numbers at the top). Further, I applied the `unlikely` changes to master (which gives a small improvement there), and diffed against that. I found that this PR is *still* a benefit, even discounting the `unlikely` effect:

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. | master Branch Misses | PR Branch Misses | -Branch Misses | master Cache Misses | PR Cache Misses | -Cache Misses | master Cache Refs | PR Cache Refs | -Cache Refs |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |    --: |    --: |   --: |
| marine_ik                        |  46616 | 295.5 | 282.0 |    4% | 909.1 | 922.4 |   -1% |  53973 |  53071 |    1% | 111557 | 111216 |    0% | 255870 | 255770 |    0% |
| mesh                             |  11306 | 307.8 | 295.0 |    4% | 984.6 | 991.9 |    0% |  10430 |  10295 |    1% |      8 |     66 | -725% |  60003 |  60853 |   -1% |
| instruments                      |   3442 | 104.4 | 100.8 |    3% | 377.4 | 374.6 |    0% |    399 |    517 |  -29% |     15 |      1 |   93% |    910 |    940 |   -3% |
| mesh.pretty                      |  24646 | 180.4 | 177.3 |    1% | 584.2 | 586.8 |    0% |  13942 |  13811 |    0% |    663 |    466 |   29% |  88013 |  88189 |    0% |
| numbers                          |   2345 | 254.2 | 251.0 |    1% | 798.1 | 802.3 |    0% |   2122 |   2138 |    0% |      0 |      0 |    0% |    322 |      5 |   98% |
| gsoc-2018                        |  51997 |  74.7 |  73.8 |    1% | 171.4 | 170.9 |    0% |  30174 |  30490 |   -1% |  55826 |  53100 |    4% | 168496 | 168384 |    0% |
| citm_catalog                     |  26987 |  86.6 |  85.8 |    0% | 319.5 | 313.4 |    1% |   1816 |   1760 |    3% |   2184 |   1150 |   47% |  89540 |  89509 |    0% |
| update-center                    |   8330 | 114.7 | 114.5 |    0% | 343.3 | 342.2 |    0% |   6032 |   5688 |    5% |      2 |      4 | -100% |  33920 |  33080 |    2% |
| random                           |   7976 | 141.9 | 142.6 |    0% | 497.0 | 493.9 |    0% |   2099 |   2081 |    0% |    332 |    205 |   38% |  37557 |  36861 |    1% |
| github_events                    |   1017 |  77.9 |  78.3 |    0% | 270.1 | 267.8 |    0% |     84 |    105 |  -25% |      0 |      0 |    0% |      1 |      0 |    0% |
| twitterescaped                   |   8787 | 189.2 | 190.4 |    0% | 511.3 | 512.3 |    0% |   4742 |   5013 |   -5% |    159 |     24 |   84% |  30050 |  29048 |    3% |
| twitter                          |   9867 |  91.5 |  92.9 |   -1% | 299.8 | 297.4 |    0% |   3125 |   3465 |  -10% |    354 |    299 |   15% |  33785 |  31769 |    5% |
| canada                           |  35172 | 289.1 | 293.8 |   -1% | 976.7 | 994.6 |   -1% |  18685 |  18518 |    0% |  22833 |  21826 |    4% | 154751 | 154738 |    0% |
| apache_builds                    |   1988 |  84.5 |  86.0 |   -1% | 308.1 | 308.1 |    0% |    104 |    112 |   -7% |      0 |      0 |    0% |      0 |      0 |    0% |